Site controller refactored
